two small fixes for C++0x conformance
git-svn-id: svn+ssh://svn.code.sf.net/p/tisch/code/libtisch@183 28952012-f250-0410-9090-e835bad21ca5

+++ libs/simplecv/Image.h
@@ -198,7 +198,7 @@ class TISCH_SHARED Image {
 		#ifndef _MSC_VER
 
 		// wrapper for semaphore access
-		inline int do_sem( int val ) {
+		inline int do_sem( short int val ) {
 			struct sembuf sem_op = { 0, val, 0 };
 			return semop( sem, &sem_op, 1 );
 		}
